requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3